requirements/base.txt
requests==2.28.2
Django==4.1.8
django-environ==0.10.0
djangorestframework==3.14.0
djangorestframework-jsonapi==6.0.0
